requirements.txt
Flask==3.1.0
llama-index-llms-groq==0.3.0
pandas==2.2.3
